Added view model for home screen, implemented fetching recipes from database instead of mocked recipes for home screen, added DI modules for home screen and recipe input screen

+++ app/src/main/java/com/example/cooke/data/mapper/SnapshotMapperImpl.kt
@@ -15,7 +15,7 @@ class SnapshotMapperImpl: SnapshotMapper{
                 imageURI = documentSnapshot["imageURI"] as String,
                 ingridients = documentSnapshot["ingridients"] as List<String>,
                 preparation = documentSnapshot["preparation"] as List<String>,
-                duration = documentSnapshot["duration"] as Float,
+                duration = documentSnapshot["duration"] as Double,
                 difficulty = documentSnapshot["difficulty"] as String,
                 isFavorite = documentSnapshot["isFavorite"] as Boolean
             )
